+++ database/migrations/2019_09_26_153342_create_devices_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateDevicesTable extends Migration
+{
+    public function up()
+    {
+        Schema::create('devices', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->index();
+            $table->string('platform')->nullable();
+            $table->string('platform_version')->nullable();
+            $table->string('browser')->nullable();
+            $table->string('browser_version')->nullable();
+            $table->boolean('is_desktop')->default(0);
+            $table->boolean('is_mobile')->default(0);
+            $table->string('language')->nullable();
+            $table->boolean('is_trusted')->default(0)->index();
+            $table->boolean('is_untrusted')->default(0)->index();
+            $table->timestamps();
+
+            // $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('devices');
+    }
+}