+++ resources/views/login-secure.blade.php
@@ -5,7 +5,7 @@
     <div class="row justify-content-center">
         <div class="col-md-8">
             <div class="card">
-                <div class="card-header">{{ __('Secure Login') }}</div>
+                <div class="card-header">{{ __('Secure login') }}</div>
 
                 <div class="card-body">
                     <form method="POST" action="{{ route('login.secure', $token) }}">